+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Please document your changes in this format:
 ```
 
 ## [Unreleased]
+
+### Added
+- Bike kitchen theme #1509 @tiltec @djahnie
+
 ### Changed
 - Rename stores to places in UI and change shopping-cart to dot-circle #1398 @tiltec @djahnie
 - Separate conversations and threads in message notifications @tiltec @djahnie